Moved PerformanceLogger class to com.unity.ml-agents/Runtime/PerformanceLogger.cs, added logging of elapsed time and steps per second, changed name of logged scalar when logging time it takes to log an episode (episodeLength --> episodeTime) to avoid ambiguity. Documented the class PerformanceLogger explaining important fields and how to use it
