--- Project/Assets/ML-Agents/Examples/3DBall/Scripts/PerformanceLogger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1778b8daa744c4e01be5c4189072da69
